refactor: 归位版本领域实现
