+++ xl_updata_tool/app/features/versions/service.py
@@ -5,14 +5,14 @@ from __future__ import annotations
 import os
 from pathlib import Path
 
-from app.core import database as db
-from app.core.local_bundle_sync import sync_local_bundles
-from app.core.seed_versions import seed_bundled_versions
-from app.core.version_cleanup import count_downloaded_bundles, delete_downloaded_bundles
-from app.core.version_data import compute_download_hashes, compute_version_delta_map
-from app.core.version_download import calculate_missing_downloads
-from app.core.version_manager import VersionManager
-from app.core.version_update import register_checked_version
+from app.platform import database as db
+from .local_bundle_sync import sync_local_bundles
+from .seed_versions import seed_bundled_versions
+from .version_cleanup import count_downloaded_bundles, delete_downloaded_bundles
+from .version_data import compute_download_hashes, compute_version_delta_map
+from .version_download import calculate_missing_downloads
+from .version_manager import VersionManager
+from .version_update import register_checked_version
 
 
 class VersionService:
