+++ xl_updata_tool/docs/worklog/issue-50-architecture-finalization/findings.md
@@ -103,3 +103,9 @@
 - `app/core/` 对应五个旧路径改为显式迁移期转发；Audio/Preview Feature 内部 import 已切到 Feature 路径，测试和预览导出调用方同步切换。
 - P6a 未删除兼容门面：`git grep` 仍能发现旧入口被测试、Shell、旧 UI 或兼容模块使用；引用归零留到 P8 的清理门。
 - P6a 新鲜验证：全量 `pytest -q -p no:cacheprovider` 通过（135 项）；Ruff 通过；不落盘 AST `AST_OK 166`；`P6_IMPORT_SMOKE_OK`；runtime Qt smoke `RUNTIME_QT_SMOKE_OK ('versions', 'preview', 'audio', 'character', 'importer')`；`git diff --check` 通过。
+
+## P6b-版本 退出复核（2026-08-24）
+
+- `app/features/versions/` 已接管 `version_cleanup`、`version_data`、`version_download`、`version_manager`、`version_update`、`local_bundle_sync` 和 `seed_versions` 的真实实现；Version Service/Controller 已切到 Feature 内部路径。
+- 对应 `app/core/version_*`、`app/core/local_bundle_sync.py`、`app/core/seed_versions.py` 均保留为迁移期兼容转发；Bundle Parser 仍暂由旧 core 兼容入口提供，留待下一批处理。
+- P6b-版本 新鲜验证：全量 `pytest -q -p no:cacheprovider` 通过（136 项）；Ruff 通过；不落盘 AST `AST_OK 173`；`VERSIONS_IMPORT_SMOKE_OK`；runtime Qt smoke `RUNTIME_QT_SMOKE_OK ('versions', 'preview', 'audio', 'character', 'importer')`；`git diff --check` 通过。
