+++ xl_updata_tool/docs/worklog/issue-50-architecture-finalization/progress.md
@@ -15,7 +15,8 @@
 - [pending] P5：迁移 Importer / AssetStudio Worker，去除 Feature 对旧导入 Worker 的生产依赖。
 - [completed] P5：迁移 Importer / AssetStudio Worker；Qt-free 导入处理进入 Feature，生产 Feature Worker 不再依赖旧 `app/ui/workers/import_as.py`，旧路径保留薄兼容门面。
 - [completed] P6a：归位音频与预览领域实现；`app/core` 对应路径降为兼容转发，Feature 内部不再从 core 获取这些领域实现。
-- [pending] P6b：归位版本与导入领域实现及其基础设施边界。
+- [completed] P6b-版本：归位版本仓库、差异、下载计划、更新持久化、本地 Bundle 同步和种子实现；对应 core 路径降为兼容转发。
+- [pending] P6b-Bundle：归位 Bundle Parser/Selector/Manager，并收口 Versions/Importer 调用边界。
 - [pending] P6c：归位角色领域仓库/展示实现。
 - [pending] P6d：收口平台与诊断基础设施。
